+++ PIDmotores/Movement.h
@@ -80,7 +80,7 @@ class Movement {
         double getFrontLeftSpeed();
         double getFrontRightSpeed();    
 
-        int getOrientation(const compass currentOrientation);
+        uint8_t getOrientation(const compass currentOrientation);
         void moveMotors(const MovementState state, const double targetOrientation);
 
         void setMotorsDirections(const MovementState state, MotorState directions[4]);
